Add files via upload
program that checks all possible combinations of a pair of numbers in the interval
of two given numbers. The output prints the combination whose sum of the numbers is equal to
a magic number. If there is no combination matching the condition is printed
message that it was not found.
Input: the input is read from the console and consists of three lines:
 First line - start of interval - integer in the range [1. . . 999]
 Second line - end of range - integer in range [greater than first number. . . 1000]
 Third line - magic number - integer in the range [1. . . 10000]
Output: one line should be printed on the console according to the result:
 If a combination is found whose sum of numbers equals the magic number
o "Combination N:{sequence number} ({first number} + {second number} = {magic number})"
 If no combination matching the condition is found
o "{the number of all combinations} combinations - neither equals {the magic number}